Separate a component for size options
