+++ src/styles/input.scss
@@ -8,5 +8,6 @@
 @use 'ProductDetailOrderForm';
 @use 'Products';
 @use 'ShoppingBag';
+@use 'SizeOptionsContainer';
 @use 'About';
 @use 'Form';
